requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
chardet==3.0.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.7
djangorestframework==3.11.1
grpcio==1.63.0
grpcio-tools==1.63.0
idna==2.10
kombu==5.3.7
param==2.1.1
pillow==10.4.0
prompt_toolkit==3.0.47
protobuf==5.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
requests==2.24.0
six==1.16.0
sqlparse==0.5.1
stability-sdk==0.8.6
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.25.11
vine==5.1.0
wcwidth==0.2.13
